+++ Toolbox/commandParser.cpp
@@ -2451,6 +2451,7 @@ int set_property(const shared_ptr<DomainBase>& domain, istringstream& command) {
 		else if(is_equal(value, "SUPERLU")) tmp_step->set_system_solver(SolverType::SUPERLU);
 		else if(is_equal(value, "MUMPS")) tmp_step->set_system_solver(SolverType::MUMPS);
 		else if(is_equal(value, "CUDA")) tmp_step->set_system_solver(SolverType::CUDA);
+		else if(is_equal(value, "PARDISO")) tmp_step->set_system_solver(SolverType::PARDISO);
 		else suanpan_error("set_property() need a valid solver id.\n");
 	} else if(is_equal(property_id, "precision")) {
 		string value;